+++ src/components/textbox/textbox-test.stories.tsx
@@ -1,8 +1,6 @@
-import React, { useState, useRef } from "react";
+import React, { useState } from "react";
 import { action } from "@storybook/addon-actions";
 import Textbox, { TextboxProps } from ".";
-import Box from "../box";
-import Button from "../button";
 import CarbonProvider from "../carbon-provider/carbon-provider.component";
 import { ICONS } from "../icon/icon-config";
 
@@ -193,42 +191,3 @@ export const PrefixWithSizes = () => {
 };
 
 PrefixWithSizes.storyName = "prefix with sizes";
-
-export const TextboxComponent = (props: Partial<TextboxProps>) => {
-  const [state, setState] = useState("Textbox");
-
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox label="Textbox" value={state} onChange={setValue} {...props} />
-  );
-};
-
-export const TextboxComponentInputRef = () => {
-  const ref = useRef(null);
-
-  return (
-    <Box margin="0 25px">
-      <Button
-        onClick={() => {
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          if (ref.current) ref.current.focus();
-        }}
-      >
-        Focus Textbox
-      </Button>
-      <TextboxComponent
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        inputRef={(el: React.ChangeEvent<HTMLInputElement>) => {
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          ref.current = el.current;
-        }}
-      />
-    </Box>
-  );
-};